Create PRCard Component
I decided to use the <Card /> component from Shad/cn so I went ahead and installed it. I added an instance of the <PRCard /> component in the <LookupPage /> and input some placeholder data so that I could style the card component itself. Once we get the API loaded and save the response data we can hook up the data and iterate over the data to create the card instances. I added icons from the Lucide library according to the Figma design.