Updates PrCard & LookupPage
In LookupPage.jsx component > loadDataFromJSON function I added logic to update the setRepo object to include the owner and repo data from uploaded data files. This will ensure the useEffect in the PRCard component correctly fetches the information for opened PRs to be displayed. I also found some unchecked events in lastEvent, like when a branch is deleted. I went ahead and added logic to correctly display that information. Line 68